Add required queries for the OPD Register by configuration class
