+++ opensrp-opd/src/main/java/org/smartregister/opd/contract/OpdRegisterFragmentContract.java
@@ -7,12 +7,9 @@ import org.json.JSONArray;
 import org.smartregister.commonregistry.CommonPersonObjectClient;
 import org.smartregister.configurableviews.model.Field;
 import org.smartregister.domain.Response;
-import org.smartregister.opd.pojos.QueryTable;
-import org.smartregister.opd.pojos.InnerJoinObject;
 import org.smartregister.view.contract.BaseRegisterFragmentContract;
 
 import java.util.List;
-import java.util.Set;
 
 /**
  * Created by Ephraim Kigamba - ekigamba@ona.io on 2019-09-13
@@ -33,15 +30,9 @@ public interface OpdRegisterFragmentContract {
 
         void updateSortAndFilter(List<Field> filterList, Field sortField);
 
-        String getMainCondition();
-
-        String getMainCondition(String tableName);
-
         String getDefaultSortQuery();
 
         String getDueFilterCondition();
-
-        String getWhereInQuery();
 
         void onClientClicked(@NonNull CommonPersonObjectClient commonPersonObjectClient);
 
@@ -50,10 +41,6 @@ public interface OpdRegisterFragmentContract {
     }
 
     interface Model {
-
-        String mainSelect(@NonNull InnerJoinObject[] tableColsInnerJoins, @NonNull QueryTable[] tableCols);
-
-        String mainSelectWhereIdsIn(@NonNull InnerJoinObject[] tableColsInnerJoins, @NonNull QueryTable[] tableCols);
 
         String getFilterText(List<Field> filterList, String filter);
