+++ opensrp-opd/src/main/java/org/smartregister/opd/model/OpdRegisterFragmentModel.java
@@ -1,18 +1,11 @@
 package org.smartregister.opd.model;
 
-
-
-import android.support.annotation.NonNull;
-
 import org.apache.commons.lang3.StringUtils;
 import org.json.JSONArray;
 import org.smartregister.configurableviews.model.Field;
 import org.smartregister.domain.Response;
 import org.smartregister.domain.ResponseStatus;
 import org.smartregister.opd.contract.OpdRegisterFragmentContract;
-import org.smartregister.opd.pojos.InnerJoinObject;
-import org.smartregister.opd.pojos.QueryTable;
-import org.smartregister.opd.utils.OpdRegisterQueryBuilder;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -24,43 +17,6 @@ import timber.log.Timber;
  */
 
 public class OpdRegisterFragmentModel implements OpdRegisterFragmentContract.Model {
-
-    @Override
-    public String mainSelect(@NonNull InnerJoinObject[] tableColsInnerJoins, @NonNull QueryTable[] tableCols) {
-        StringBuilder query = new StringBuilder();
-
-        for (int i = 0; i < tableColsInnerJoins.length; i++) {
-            InnerJoinObject tableColInnerJoin = tableColsInnerJoins[i];
-
-            OpdRegisterQueryBuilder countQueryBuilder = new OpdRegisterQueryBuilder();
-            countQueryBuilder.SelectInitiateMainTable(tableColInnerJoin);
-            countQueryBuilder.mainCondition(tableColInnerJoin.getMainCondition());
-
-            if (i != 0) {
-                query.append(" UNION ALL ");
-            }
-
-            query.append(countQueryBuilder.getSelectquery());
-        }
-
-        for (int i = 0; i < tableCols.length; i++) {
-            QueryTable tableCol = tableCols[i];
-
-            OpdRegisterQueryBuilder countQueryBuilder = new OpdRegisterQueryBuilder();
-            countQueryBuilder.SelectInitiateMainTable(tableCol.getTableName(), tableCol.getColNames());
-            countQueryBuilder.mainCondition(tableCol.getMainCondition());
-
-            if (query.length() != 0 || i != 0) {
-                query.append(" UNION ALL ");
-            }
-
-            query.append(countQueryBuilder.getSelectquery());
-        }
-
-        return query.toString();
-    }
-
-
 
     @Override
     public String getFilterText(List<Field> list, String filterTitle) {
@@ -101,67 +57,4 @@ public class OpdRegisterFragmentModel implements OpdRegisterFragmentContract.Mod
         return null;
     }
 
-    @Override
-    public String mainSelectWhereIdsIn(@NonNull InnerJoinObject[] tableColsInnerJoins, @NonNull QueryTable[] tableCols) {
-        StringBuilder query = new StringBuilder();
-
-        for (int i = 0; i < tableColsInnerJoins.length; i++) {
-            InnerJoinObject tableColInnerJoin = tableColsInnerJoins[i];
-
-            OpdRegisterQueryBuilder countQueryBuilder = new OpdRegisterQueryBuilder();
-            countQueryBuilder.SelectInitiateMainTable(tableColInnerJoin);
-            countQueryBuilder.mainCondition(tableColInnerJoin.getMainCondition());
-
-            String idCol = "_id";
-            if (countQueryBuilder.getSelectquery().contains("JOIN")) {
-                idCol = tableColInnerJoin.getFirstTable().getTableName() + ".id";
-            }
-
-            if (countQueryBuilder.getSelectquery().contains("WHERE")) {
-                countQueryBuilder.addCondition(" AND ");
-
-            } else {
-                countQueryBuilder.addCondition(" WHERE ");
-            }
-            countQueryBuilder.addCondition("%s IN (%s)");
-            countQueryBuilder.setSelectquery(countQueryBuilder.getSelectquery().replaceFirst("%s", idCol));
-
-            if (i != 0) {
-                query.append(" UNION ALL ");
-            }
-
-            query.append(countQueryBuilder.getSelectquery());
-        }
-
-
-        for (int i = 0; i < tableCols.length; i++) {
-            QueryTable tableCol = tableCols[i];
-
-            OpdRegisterQueryBuilder countQueryBuilder = new OpdRegisterQueryBuilder();
-            countQueryBuilder.SelectInitiateMainTable(tableCol.getTableName(), tableCol.getColNames());
-            countQueryBuilder.mainCondition(tableCol.getMainCondition());
-
-            String idCol = "_id";
-            if (countQueryBuilder.getSelectquery().contains("JOIN")) {
-                idCol = tableCol.getTableName() + ".id";
-            }
-
-            if (countQueryBuilder.getSelectquery().contains("WHERE")) {
-                countQueryBuilder.addCondition(" AND ");
-
-            } else {
-                countQueryBuilder.addCondition(" WHERE ");
-            }
-            countQueryBuilder.addCondition("%s IN (%s)");
-            countQueryBuilder.setSelectquery(countQueryBuilder.getSelectquery().replaceFirst("%s", idCol));
-
-            if (query.length() != 0 || i != 0) {
-                query.append(" UNION ALL ");
-            }
-
-            query.append(countQueryBuilder.getSelectquery());
-        }
-
-        return query.toString();
-    }
 }
